+++ tests/dark-mode.test.js
@@ -17,7 +17,7 @@ crosscheck(() => {
     return run(input, config).then((result) => {
       expect(result.css).toMatchFormattedCss(css`
         ${defaults}
-        :is(.dark .dark\:font-bold) {
+        :is(:where(.dark) .dark\:font-bold) {
           font-weight: 700;
         }
       `)
@@ -40,7 +40,7 @@ crosscheck(() => {
     return run(input, config).then((result) => {
       expect(result.css).toMatchFormattedCss(css`
         ${defaults}
-        :is(.test-dark .dark\:font-bold) {
+        :is(:where(.test-dark) .dark\:font-bold) {
           font-weight: 700;
         }
       `)