+++ tests/format-variant-selector.test.js
@@ -348,8 +348,8 @@ crosscheck(() => {
       ${'.parent::before &:hover'}                             | ${'.parent &:hover::before'}
       ${':where(&::before) :is(h1, h2, h3, h4)'}               | ${':where(&) :is(h1, h2, h3, h4)::before'}
       ${':where(&::file-selector-button) :is(h1, h2, h3, h4)'} | ${':where(&::file-selector-button) :is(h1, h2, h3, h4)'}
-      ${'#app :is(.dark &::before)'}                           | ${'#app :is(.dark &)::before'}
-      ${'#app :is(:is(.dark &)::before)'}                      | ${'#app :is(:is(.dark &))::before'}
+      ${'#app :is(:where(.dark) &::before)'}                           | ${'#app :is(:where(.dark) &)::before'}
+      ${'#app :is(:is(:where(.dark) &)::before)'}                      | ${'#app :is(:is(:where(.dark) &))::before'}
       ${'#app :is(.foo::file-selector-button)'}                | ${'#app :is(.foo)::file-selector-button'}
       ${'#app :is(.foo::-webkit-progress-bar)'}                | ${'#app :is(.foo)::-webkit-progress-bar'}
       ${'.parent::marker li'}                                  | ${'.parent li::marker'}