+++ tests/kitchen-sink.test.js
@@ -305,7 +305,7 @@ crosscheck(({ stable, oxide }) => {
         }
         .drop-empty-rules:hover,
         .group:hover .apply-group,
-        :is(.dark .apply-dark-mode) {
+        :is(:where(.dark) .apply-dark-mode) {
           font-weight: 700;
         }
         .apply-with-existing:hover {
@@ -340,7 +340,7 @@ crosscheck(({ stable, oxide }) => {
         .apply-order-b {
           margin: 1.5rem 1.25rem 1.25rem;
         }
-        :is(.dark .group:hover .apply-dark-group-example-a) {
+        :is(:where(.dark) .group:hover .apply-dark-group-example-a) {
           --tw-bg-opacity: 1;
           background-color: rgb(34 197 94 / var(--tw-bg-opacity));
         }
@@ -741,9 +741,6 @@ crosscheck(({ stable, oxide }) => {
             transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
           }
         }
-        :is(.dark .dark\:custom-util) {
-          background: #abcdef;
-        }
         @media (min-width: 640px) {
           .sm\:text-center {
             text-align: center;
@@ -788,9 +785,6 @@ crosscheck(({ stable, oxide }) => {
               transition-duration: 0.15s;
               transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
             }
-            :is(.dark .md\:dark\:motion-safe\:foo\:active\:custom-util:active) {
-              background: #abcdef !important;
-            }
           }
           @media (min-width: 640px) {
             .md\:sm\:text-center {
@@ -806,6 +800,16 @@ crosscheck(({ stable, oxide }) => {
         @media (min-width: 640px) and (max-width: 767px), (max-width: 868px) {
           .multi\:text-left {
             text-align: left;
+          }
+        }
+        :is(:where(.dark) .dark\:custom-util) {
+          background: #abcdef;
+        }
+        @media (min-width: 768px) {
+          @media (prefers-reduced-motion: no-preference) {
+            :is(:where(.dark) .md\:dark\:motion-safe\:foo\:active\:custom-util:active) {
+              background: #abcdef !important;
+            }
           }
         }
       `)
@@ -874,7 +878,7 @@ crosscheck(({ stable, oxide }) => {
         }
         .drop-empty-rules:hover,
         .group:hover .apply-group,
-        :is(.dark .apply-dark-mode) {
+        :is(:where(.dark) .apply-dark-mode) {
           font-weight: 700;
         }
         .apply-with-existing:hover {
@@ -908,7 +912,7 @@ crosscheck(({ stable, oxide }) => {
         .apply-order-b {
           margin: 1.5rem 1.25rem 1.25rem;
         }
-        :is(.dark .group:hover .apply-dark-group-example-a) {
+        :is(:where(.dark) .group:hover .apply-dark-group-example-a) {
           background-color: #22c55e;
         }
         @media (min-width: 640px) {
@@ -1299,9 +1303,6 @@ crosscheck(({ stable, oxide }) => {
             transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
           }
         }
-        :is(.dark .dark\:custom-util) {
-          background: #abcdef;
-        }
         @media (min-width: 640px) {
           .sm\:text-center {
             text-align: center;
@@ -1346,9 +1347,6 @@ crosscheck(({ stable, oxide }) => {
               transition-duration: 0.15s;
               transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
             }
-            :is(.dark .md\:dark\:motion-safe\:foo\:active\:custom-util:active) {
-              background: #abcdef !important;
-            }
           }
           @media (min-width: 640px) {
             .md\:sm\:text-center {
@@ -1364,6 +1362,15 @@ crosscheck(({ stable, oxide }) => {
         @media (min-width: 640px) and (max-width: 767px), (max-width: 868px) {
           .multi\:text-left {
             text-align: left;
+          }
+        }
+        :is(:where(.dark) .dark\:custom-util) {
+          background: #abcdef;
+        }
+        @media (min-width: 768px) {
+          @media (prefers-reduced-motion: no-preference) {
+          :is(:where(.dark) .md\:dark\:motion-safe\:foo\:active\:custom-util:active) {
+            background: #abcdef !important;
           }
         }
       `)