+++ tests/variants.test.js
@@ -1172,7 +1172,7 @@ crosscheck(({ stable, oxide }) => {
     let result = await run(input, config)
 
     expect(result.css).toMatchFormattedCss(css`
-      :is(.dark :is([dir='rtl'] .dark\:rtl\:italic)) {
+      :is(:where(.dark) :is(:where([dir='rtl']) .dark\:rtl\:italic)) {
         font-style: italic;
       }
     `)
@@ -1196,7 +1196,7 @@ crosscheck(({ stable, oxide }) => {
     let result = await run(input, config)
 
     expect(result.css).toMatchFormattedCss(css`
-      :is(.dark :is([dir='rtl'] .dark\:rtl\:placeholder\:italic))::placeholder {
+      :is(:where(.dark) :is(:where([dir='rtl']) .dark\:rtl\:placeholder\:italic))::placeholder {
         font-style: italic;
       }
     `)